// infrastructure/lib/config/types.d.ts
export interface EnvironmentConfig {
    env: Environment;
    environmentName: EnvironmentName;
    network: NetworkConfig;
    database: DatabaseConfig;
    redis: RedisConfig;
    ecs: EcsConfig;
    amplify: AmplifyConfig;
    security: SecurityConfig;
    monitoring: MonitoringConfig;
    costOptimization: CostOptimizationConfig;
    compliance: ComplianceConfig;
    features: {
        enableCdn: boolean;
        enableCaching: boolean;
        enableElasticsearch: boolean;
        enableEventBridge: boolean;
        enableStepFunctions: boolean;
        enableApiGateway: boolean;
        enableCognito: boolean;
        enableSes: boolean;
        enableS3: boolean;
        enableCloudFront: boolean;
        enableRoute53: boolean;
        enableCertificateManager: boolean;
        enableSecretsManager: boolean;
        enableParameterStore: boolean;
        enableEventSourcing: boolean;
        enableMessageQueuing: boolean;
        enableStreamProcessing: boolean;
        enableMachineLearning: boolean;
        enableAnalytics: boolean;
        enableDataLake: boolean;
    };
    tags: {
        [key: string]: string;
    };
}
export interface StackProps {
    env: Environment;
    config: EnvironmentConfig;
    description?: string;
    tags?: {
        [key: string]: string;
    };
}
export interface DatabaseStackProps extends StackProps {
    vpc: any;
    databaseSecurityGroup: any;
    redisSecurityGroup: any;
}
export interface BackendStackProps extends StackProps {
    vpc: any;
    cluster: any;
    loadBalancer: any;
    backendSecurityGroup: any;
    database: any;
    redis: any;
    assetsBucket: any;
}
export interface FrontendStackProps extends StackProps {
    backendUrl: string;
}